Do not use dynamic titles in terminals
Dynamic titles make it inherently difficult to operate on windows with
xdotool/wmctrl, as the command needs the window title to operate on, but
once executed, the term's title changes to something *including the
title*. So one ends up with two windows with the same string in the
title, making wmctrl/xdotool operate on the first one, which is the term
you originally started out at.
This is not what we want.